Add config update via HTTP endpoint, enforced retention on DB close
This commit introduces a new HTTP endpoint that could be started
by specifying the `config-addr` field. This will launch an endpoint
at `config-addr`/config and listen to GET requests. A GET request
will trigger a config reload from disk (you do not send the new
config in that request, you simply signal that it has changed).

The main litestream process will then replace its old config with the
new one, add new databases to replication and remove those that are
being replicated, but are missing from the new config. Other database
connections are being kept as-is, which was not the case before - prior
to this change, the whole litestream process had to be restarted. Now,
the goroutines for databases that are kept keep running, goroutines
for old databases are safely stopped, and new ones are started.

In coupe with the new option called `enforce-retention-on-close`,
litestream will enforce retention when a database is removed from the
config. The retention count is hardcoded to be 0 in this case, which
leaves us with only one snapshot and one WAL file. This could be
further improved to remove the snapshots completely, leaving a single
database file in the blob storage, but that can be also done in
server-space.